Implement basic three phase commit
This commit is an unusually large one, but attempts to prove out a basic
architecture for doing a fast single-threaded three-phase commit.  It
introduces data structures for library consumers to populate and to
perform actions from.  This is a very rudimentary first pass, tests are
only partially populated, and there are big pieces of missing function.
Notably:

  No epoch change
  No checkpoints
  No watermark advancement
  No documentation
  No inter-node integration testing

What this commit does include is.

  Single node mock integration test for commit of a single seqno

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>