Add basic sample and integration tests
This introduces a rudimentary serial consumer of the mirbft library.
The sample requires that the user supply a hasher, a validator, and a
network link.

This sample is then used in the top level integration test to commit
1000 messages with a single node, and with a network of 4 nodes.

Note, all of the deficiencies noted in the previous commit are still
present.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>