Factor out the proposal logic from buckets
Ultimately, whether data belongs in a bucket or not does not matter.
We have bucket assignment and rotation for censorship resistance, so
long as the message is sent to the correct leader, and the leader for a
particular message rotates, we are good.  This commit moves 'propose' to
be a simple round robin across all buckets owned by a node.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>