Intermediate commit for implementing epoch-change
This commit adds a fair bit of code that is never exercised in any
integration (or unit) test.  It doesn't work yet and contains
placeholders, but doesn't break any of the green-path tests.  Committing
simply to avoid screwing something up and not being able to revert to a
good point.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>