Fix flakiness in main test
This commit consolidates the logic of the single node and multinode
tests into a single table tests.  It also adds a third test case for a 3
node CFT setup.  And finally, it changes the pending epoch state to
continue to broadcast the epoch change messages until the new epoch is
received (or, in the future times out).

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>