Remove bucket from external messages
This is part of a significant refactor which will treat sequences by
'column and bucket' within an epoch, but as vanilla sequences outside.
This hopefully will allow us the best of both worlds, being able to
enforce the natural prepare/commit order within an epoch, while avoiding
the complexity of differing epoch configurations on number of buckets
during epoch change.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>