Begin migrating to literal seqno for entry
As the code stands at the moment, the *sequence structure holds a
reference to an entry, with a 'SeqNo', that is actually a column number.
This creates some annoying translation back and forth, and is error
prone.  This is the first commit in a small journey to remove all the
mangling.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>