Untangle epoch change and epoch
For expediency, epoch change was originally implemented as part of the
state of the epoch, but it doesn't really make sense there.  The epoch
certainly has a lifecycle, but we need to collect suspicions and epoch
changes across multiple epoch numbers concurrently, and the whole
three-phase new-epoch commit is a bit of a distraction in the green
path.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>