Fix epoch change to trigger on lack of progress
Presently, epoch change only truly works for the initial epoch change
where there are no entries in the pSet or qSet.  This commit gets the
process wired through correctly to the point that it fails, as there is
currently no way to retrieve the requests which preprepared in a
previous epoch.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>