Step towards persistence
This commit does several things.  First, it introduces the notions of
the PEntries and QEntries.  Entries that will ultimately need to be
persisted prior to sending messages in order to be crash safe.

Secondly, it reorders the state machine for the sequence.  Whereas
before we sent the preprepare before computing a digest over it, because
having the digest available before creating the QEntry is useful, we now
digest beforehand.  We also consolidate digest and validation by
including a source for the batch.  The application may choose to skip
the validation if the source is itself.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>