Move proposer out of epoch
The main barrier to enabling the epoch change test was the fact that
proposals mid-preprocess could be lost.  This was because if there was
no active epoch when the preprocess result arrived, there was no
proposer in which to queue the proposal for later inclusion in a batch.

This commit moves the proposer up to the state machine, and injects it
as a dependency of the epoch.  Then, when the epoch has room (in any
bucket), it either forwards the proposal to the correct node, or
allocates a sequence for it.

This is still not as robust as I'd like it, there is still a risk for
loss when bucket rotation is happening (which is not yet implemented).
Additionally, the test runs _very_ slowly (8 seconds), because we must
epoch change for each and every log window, since nodes cannot be
excluded as of yet in the new epoch.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>