Capture panics and exit status more sanely
The existing behavior of the statemachine was that on panic, it would
try to print its status before continuing to panic.  This generally
violates the principal that libraries should not panic, and does not
give the user the opportunity to capture this status and for instance,
transform it to JSON instead of pretty printing it to the console.  It's
still expected that users generally panic when the state machine errors.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>