Add first incarnation of testengine
The current mirbft tests are done using a plurality of goroutines, being
randomly scheduled according to the go runtime.  This creates a few
problems:

1) Despite having so many go routines, because they are all competing
for resources, the tests actually run rather slowly.
2) Although Mir is tick based, and these go routines may inject ticks at
a much faster rate of time than a system practically would (say, every
10ms, instead of every 500ms), the reality is that 10ms is a lot of
wasted time, which may happen hundreds or thousands of times in a test.
3) When failures occur, they're not easily reproducible.

Because of this, (and known since the projects inception), we need a
deterministic, time-free environment to run tests in.  This commit is
the first step down that path.  It adds an event log which is
serializable and contains all possible operations performed on a mirbft
node.  It includes a 'Player' which can apply that log to a set of
nodes, and a 'Recorder' which  appends to the log based on the node
configuration.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>