Add way to propose non-blocking, testengine client
Having a client appropriately track whether a particular replica has
advanced this client's watermarks is a challenging proposition.  So, by
default the API exposes a blocking 'Propose' which will simply wait
until the client's request enters the watermarks (or the context
expires).  However, when trying to implement a deterministic testengine,
this is a problem, and further, some clients may wish to be notified
immediately if their request falls outside of watermarks, rather than
waiting on a timeout.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>