Cleanup testengine common code
Most tests will want to start with the same sort of basic testengine
configuration, so requiring that everyone implement it is unnecessary.
This commit creates a new 'BasicRecorder' function, which given a number
of nodes, clients, and requests per client, bootstraps a valid recorder.
The caller may then tweak the recorder before creating a recording based
on that recorder.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>