Add deterministic silencing test, implement shrink
Up until now, the epoch change has left the leader set alone.  This CR
adds the actual logic to reduce the leader set, it functions roughly as
follows:

If the old leader set had only one node, select yourself.

If the lastActiveEpoch was the previous epoch, try to identity the
lowest non-prepared sequence number.  If there is one, mark the leader
of the bucket corresponding to that seqno as bad, and remove it from the
new leader set.

If the bad node cannot be identified, or the last active epoch was not
the previous epoch number, eject the previous epoch leader.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>