Move self-messaging from recorder to player
In order to replay the log, since self-message is done immediately at
process, and not as a recv event, we need to have this processing done
in the player, rather than in the recorder.  Otherwise, playing back a
recording will not get us to the same state.  This commit also adds a
test which verifies that a recording can be played back to the same
state.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>