Add Mangling helper to make human readable tests
The mangling structure is very simple to evaluate from a machine
perspective.  However, it's not always obvious, and can be verbose to
construct nested manglers to accomplish some specific task.

This commit introduces a human friendly syntax for specifying test
manglings.  For instance:

 testengine.Drop().AtPercent(3).Messages().FromNodes(2,7)

will construct a mangler which will drop 3% of the messages which
originate from nodes 2 or 7.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>