Track correct client reqs, don't poll
Prior to this commit, the sequences would basically look up the client
requests every time its status might have changed, and if all clients
were present then proceed to the next state.  This was ugly and
expensive.  This commit begins to disassociate the sequence from the
client windows by having the outstanding requests track which requests
are outstanding and consuming a stream of them from the client windows
as they become available as correct.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>