Track commits at top level of state machine
Prior to this commit, both the epoch changer and the epoch itself would
generate commits and apply them to update the client windows.  This was
duplicated code and obfuscated the fact that the client windows really
depend only on the committed state, not on the PBFT type state.  This
commit moves that logic to a top level advance loop in the state machine
which will be used to consolidate other logic like watermark movement in
the future.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>