Begin converting ActionResults to event
The action results currently uses a public member to carry the
association between the hash results and the original hash request.
This has always been a little bit awkward and bloats the user facing
API.  As part of a transition towards a pure event based recordable
model for the state machine, transitioning this structure to using a
member which holds a proto structured form of the event, which will
eventually be a private member.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>