Report actions received events
This is a bit of a weird event for the state machine to process, in
that, it does not impact the state (in much the same way that the
propose does not actually impact the state), but it also does not return
any actions.

However, when recording inputs to the state machine, and attempting to
replay them, it is nice/necessary to associate which set of Actions
corresponds with which set of ActionResults.  We've seen this in the
testengine implementation, and as we look towards a more general state
machine log capture mechanisms for less orchestrated testing and
debugging, it will be useful there as well.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>