Properly count iterations of the DrainClients loop
Previously, we had relied on the fact that the event log retained events
even after they had been processed.  Since that is no longer the case,
the Count() from the event log was always small and tests never timed
out.  This change uses the existing count variable in the DrainClients
loop to break on instead and removes the pointless method in the
eventlog.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>