Move client windows to a committed mask
The current client windows state specifies the low watermark for each
bucket.  This works well enough in the green, non-config path, but, it's
actually possible given more interesting byzantine behavior, and, when
the number of buckets change, to develop holes in the client windows for
a given bucket.

The approach taken here is to represent all of the committed request
numbers beyond the low watermark at a bitmask, relative to the low
watermark.  In the green path, this bitmask is empty, only when there
are holes does this bitmask begin to be populated, such as when one of
the leaders crashes.  This is generally still okay, as the bitmask
representation is still relatively concise, and for most clients, the
size is still small, and once the epoch change concludes, the bitmasks
quickly return to their original, near-zero size.

Additionally, this change gives us the ability to specify a client width
(something that was a TODO anyway, but was necessary to infer the zero
values of a partial bitmask).

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>