Gzip the event logs
This change I was a bit on the fence about.  In general, the event logs
contain a lot of very repetitive information, such as enumerating the
same hash many times.  So, gzip actually does a fairly good job
compressing the logs out of the box -- my initial testing was around 75%
reduction in size.

This makes working with the logs a bit more of a pain, but also gets us
better error handling to deal with truncated or malformed logs, so, all
in all, seems worth doing.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>