Move status to sub-package
I hate the idea of creating a sub-package as a simple directory, but
having all of the assorted status types and enumerations in the primary
godoc was very distracting and difficult to grasp that it was all purely
status reporting related.

Related to this, the status enums were exported for types (or not), and
re-used in the status reporting.  These have been split and duplicated
(we should really see about adding a test to make sure they stay in
sync).

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>