Begin tracking multiple correct requests
The current client tracker code is rather naive and does not handle the
case where clients shut down improperly, or byzantinely or accidentally
submit multiple requests for the same request number.  There is also no
retransmission logic, fetch state, etc.  This is the first step to
bringing that into place.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>