Stop persisting epoch config with checkpoint
This never made a lot of sense, especially as the epoch configuration
could have already changed by the time the checkpoint comitted.
We were taking care to ensure that the epoch configuration under which
the sequence committed is what was persisted, but actually, we need the
active epoch in the log to be what's there, because after truncation,
we'll need to appropriately assign the epoch to the persisted
pEntry/qEntry items.

Not tackled in this commit is soliciting the active epoch configuration.
This is something that very likely will need to be done after state
transfer, or at startup, but we'll tackle it when we get there.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>