Begin commit based state movement
This is the beginning of a refactor which will transition the state
machine towards using a particular committed configuration for
validation of client request and batch contents.  It will take a few
commits to accomplish this... so this one will not compile and is
generally broken.  I hate to break bisection, but, doing this all in one
commit is a bit risky.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>