Initial wiring of commit based network state
This concludes the series of WIPs (which are left as is for posterity)
which begins allocating sequences in the epoch, as well as watermarks in
the client, based off of the currently committed network state.

There was a logical bug in the previous code which had assumed that if
the watermarks of the client tracker and of the epoch moved in unison,
that we did not need to track what requests were eligible for which
sequence number.  But of course this isn't actually true, because the
movement is not uniform across replicas.  In this case, it's possible
that someone will have moved to a stable checkpoint c2, while the other
replica is still generating c2.  In this case, the lagging replica will
disagree that the watermarks have moved as far as the first replica
believes.  In fact, just like with all reconfiguration in order to
prevent halting at checkpoints, we rely on the commit state from the
_previous_ checkpoint in all cases, to ensure that any replica
processing the batch has the proper configuration to validate it.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>