Overhaul mangling framework
The previous mangling framework was very simple, but lacked context when
writing new manglers, so logic ended up being duplicated repeatedly.
The new mangling framework is quite a bit less readable, as it depends
on reflection and a sort of 'mix-in' of mangling components, but also
requires significiantly less code and is easier to extend.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>