Introduce allocated requests
This is a long-coming change to make the client APIs a bit more sane.
The state machine explicitly allocates request numbers to clients which
the consumer may then send back requests for.  This eliminates the
requirement to inspect network state via checkpoint requests to decide
when it's safe to inject new client requests.  It will also remove the
need for the clientWaiter hack, which has always been smelly.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>