Split out client actions.
Historically, client actions have been comingled with standard node
actions.  This presents a number of dificulties/annoyances.
Particularly, the application needs to know what requests clients are
allowed to submit, but this state is potentially changing constantly.
In order for these to coordinate, we will need to tolerate client
locking and generally deal with problems that do not exist in the
vanilla state-machine -> WAL -> network path.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>