Add client actions received
Up until this point, the client proposer has been simply calling into
the node "Propose" API.  But this is really a poor way to deal with
clients.  Instead, we want the client actions to run on their own
thread, aggregating results just like the regular actions thread, and
periodically call back into the state machine with batches of requests.
This should produce far fewer events, and therefore have less locking
contention in the serializer.

This additionally removes the bespoke action printing in mircat as the
state machine now returns actions as a text serializable proto.  There
appears to be some printing of unset fields that we'll need to look
into, but overall it cleans up the code and increases readability.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>