Enable real-ish client semantics in test.
For a while in this commit series, we had stopped dynamically injecint
client requests and instead was storing them a priori in the DB to be
found by the client processor.  This commit introduces logic in the
client processor to properly build up client requests to inject into the
system when they arrive after the client request number has been
allocated.  It then modifies the tests to take advantage of this fact.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>