Convert actions to stream/list
Today we have a structure with repeated fields, which does not preserve
order across fields.  This was done originally for ease of batching
items together, but has created problems of its own.  Instead, we can
pushing the batching to the edges, this will make the state logs more
readable and give more flexibility on the batching boundaries.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>