Rework testengine to use processors
The testengine started out simple, but gained more and more bespoke
action processing that has created two parallel paths, resulting in dual
maintenance and a lot of testengine code that is being tested instead of
the production path.

This change removes a large amount of test code, but also breaks the
mangler framework so currently, many tests are passing on the green path
when they should be testing a red path.  To be fixed before being pushed
upstream.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>