Refactor top-level structure - first pass (#33)
Refactor top-level structure - first pass

First pass of bringing more order and structure to the code at the top level.
The main change is the addition of comments to properly document the code.

Files with "Refactored: 1" in the initial comment can be considered
cleaned up. They still contain plenty of TODOs, but they should be on the
right path. Other files are subject to potentially big changes.

Signed-off-by: Matej Pavlovic <mpa@zurich.ibm.com>

Co-authored-by: Matej Pavlovic <mpa@zurich.ibm.com>