Remove Actions and Results, keep only Events
This commit unifies the types of items passed between the modules.
Before, the state machine would consume Events and produce actions,
and other modules would consume actions and produce results, which were
passed back to the state machine as events. Moreover, there were some
exceptions to this basic rule and the whole thing became complicated to
understand.

Now everything is just Events. The modules consume Events at the input
and produce events on their output, and the routing happens in a simpler
way, only based on the Event type.

Signed-off-by: matejpavlovic <matopavlovic@gmail.com>