Change Net's reception interface, add gRPC impl.
The Net module's reception interface now works with a channel.
The ReceiveChan() function returns a channel, to which the module
writes all incoming messages. This should be simpler to both implement
and to use. It also probably makes it simpler to have multiple threads
read incoming messages.

A simple implementation of the Net module that uses the real network
for sending messages is now in pkg/grpctransport. It is based on gRPC,
but does not use any authentication (yet) and needs to be initialized
with static membership.

Signed-off-by: matejpavlovic <matopavlovic@gmail.com>