Add network request receiver and dummy client
The RequestReceiver is a component external to the Node implemented with gRPC
that can receive requests over the network and submit it to the Node.
The DummyClient is a simple gRPC client that sends requests as gRPC requests
to the Nodes' request receivers.
Everything is currently based on pre-configured static membership.

Signed-off-by: matejpavlovic <matopavlovic@gmail.com>