Fix Protobuf code generation incompatibility
The go files generated by the protobuf compiler used the
github.com/golang/protobuf/proto package, whereas some
(hand-written) go file used google.golang.org/protobuf/proto.
All is now unified in favor of the former. As the latter
seemst to be the newer one, this should be changed in the future
such that all files (also the generated ones) use the latter.

Signed-off-by: matejpavlovic <matopavlovic@gmail.com>