Crypto module and signing and verifying events
The Crypto module implementation can be used raw and have the
protocol add the keys based on dynamic membership. This is not
yet implemented, however.
For testing purposes, the Dummy crypto implementation can
be used - it does not perform any cryptographic operations
or key management whatsoever.
The full implementation can also be instantiated in a
pseudo-random manner, generating the same set of keys on every
node of a distributed deployment, if given the same seed. This is
useful for testing static membership deployments.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>