Use proper epoch numbers with checkpoint messages
The epoch number of a checkpint now always corresponds to the epoch
the associated sequence number belongs to.
At epoch transition, rathen than having an epoch "end" with a checkpoint,
the next epoch "starts" with a checkpoint, and thus the checkpoint
is associated with the new epoch rather than with the old one.
This also makes more sense with respect to the associated sequence number.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>