Consolidate event processing, fix data races
The processor would dispatch the same list of events to the workers
and to the interceptor. The workers were modifying the events
(by stripping off the follow-ups) concurrently read by the interceptor,
resulting in corrupted event logs produced by the interceptor.

Now it is the workers that submit the events to the interceptor
after (and not concurrently with) stripping off the follow-ups.
The code of the workers is now significantly de-duplicated and simplified.

Another minor data race in the request receiver is also fixed.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>