Improve event processing and polish mircat
* Fix data races between event processing and interception

The processor would dispatch the same list of events to the workers
and to the interceptor. The workers were modifying the events
(by stripping off the follow-ups) concurrently read by the interceptor,
resulting in corrupted event logs produced by the interceptor.

Now it is the workers that submit the events to the interceptor
after (and not concurrently with) stripping off the follow-ups.

Another minor data race in the request receiver is also fixed.

* Event processing code

The code of the workers is now de-duplicated and significantly simplified.

* Polishing mircat

The mircat code is now a bit reorganized and simplified.