Bug in CommitLogEntry hashing, view change
A direct pointer to a CommitLogEntry was used in a HashOrigin
when requesting to hash the former. When the HashResult arrived,
the digest was added to the CommitLogEntry concurrently with the
HashRequest (pointing to the CommitLogEntry) being intercepted
and stored, resulting in a data race. Now the preliminary
CommitLogEntry (that is missing the hash) is stored locally by
the ISS protocol in a map and only the map's key (the
corresponding sequence number) is used as HashOrigin.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>